+++ ta4j-core/src/test/java/org/ta4j/core/indicators/UlcerIndexIndicatorTest.java
@@ -57,24 +57,22 @@ public class UlcerIndexIndicatorTest extends AbstractIndicatorTest<Indicator<Num
 
         assertNumEquals(0, ulcer.getValue(0));
 
-        // From:
-        // http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:ulcer_index
-        assertNumEquals(1.3047, ulcer.getValue(26));
-        assertNumEquals(1.3022, ulcer.getValue(27));
-        assertNumEquals(1.2156, ulcer.getValue(28));
-        assertNumEquals(0.9967, ulcer.getValue(29));
-        assertNumEquals(0.7257, ulcer.getValue(30));
-        assertNumEquals(0.453, ulcer.getValue(31));
-        assertNumEquals(0.4284, ulcer.getValue(32));
-        assertNumEquals(0.4284, ulcer.getValue(33));
-        assertNumEquals(0.4284, ulcer.getValue(34));
-        assertNumEquals(0.4287, ulcer.getValue(35));
-        assertNumEquals(0.5089, ulcer.getValue(36));
-        assertNumEquals(0.6673, ulcer.getValue(37));
-        assertNumEquals(0.9914, ulcer.getValue(38));
-        assertNumEquals(1.0921, ulcer.getValue(39));
-        assertNumEquals(1.3161, ulcer.getValue(40));
-        assertNumEquals(1.5632, ulcer.getValue(41));
-        assertNumEquals(1.7609, ulcer.getValue(42));
+        assertNumEquals(0, ulcer.getValue(0));
+        assertNumEquals(1.2340096463740846, ulcer.getValue(26));
+        assertNumEquals(0.560553282860879, ulcer.getValue(27));
+        assertNumEquals(0.39324888828140886, ulcer.getValue(28));
+        assertNumEquals(0.38716275079310825, ulcer.getValue(29));
+        assertNumEquals(0.3889794194862251, ulcer.getValue(30));
+        assertNumEquals(0.4114481689096125, ulcer.getValue(31));
+        assertNumEquals(0.42841008722557894, ulcer.getValue(32));
+        assertNumEquals(0.42841008722557894, ulcer.getValue(33));
+        assertNumEquals(0.3121617589229034, ulcer.getValue(34));
+        assertNumEquals(0.2464924497436544, ulcer.getValue(35));
+        assertNumEquals(0.4089008481549337, ulcer.getValue(36));
+        assertNumEquals(0.667264629592715, ulcer.getValue(37));
+        assertNumEquals(0.9913518177402276, ulcer.getValue(38));
+        assertNumEquals(1.0921325741850083, ulcer.getValue(39));
+        assertNumEquals(1.3156949266800984, ulcer.getValue(40));
+        assertNumEquals(1.5606676136361992, ulcer.getValue(41));
     }
 }